requirements.txt
# Core dependencies
requests>=2.28.0
speedtest-cli>=2.1.3
pyyaml>=6.0
psutil>=5.9.0
python-dotenv>=0.21.0

# API server
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=2.0.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Optional for visualization
matplotlib>=3.5.0
pandas>=1.5.0
numpy>=1.24.0